feat: crawler pipeline
- add Crawler concrete implementation
